Voting screen to globally accessible
