+++ common/src/main/res/layout/fragment_votes.xml
