+++ common/src/main/res/layout/fragment_votes_entry.xml
