+++ common/src/main/res/layout/fragment_votes_tab.xml
