+++ common/src/main/res/values/strings.xml
@@ -18,4 +18,13 @@
     <string name="stored"><![CDATA[<b>Stored:</b> %s]]></string>
 
     <string name="x_bytes">%d B</string>
+
+    <string name="vote">Vote</string>
+    <string name="voteUser1">Vote User 1</string>
+    <string name="voteUser2">Vote User 2</string>
+    <string name="bounty_payout">Pay %1$s from our common funds to artist named %2$s?</string>
+    <string name="bounty_payout_message">Do you agree on paying %1$s from our common funds to artist named %2$s?\nThere are already %3$d upvotes, %4$d downvotes, and %5$d" undecided votes</string>
+    <string name="bounty_payout_upvoted">You voted in favor of paying %1$s from our common funds to artist named %2$s</string>
+    <string name="bounty_payout_downvoted">You voted against paying %1$s from our common funds to artist named %2$s</string>
+    <string name="bounty_payout_all_voted">Everyone voted, based on the majority we decided to pay the artist</string>
 </resources>
