+++ currencyii/src/main/res/values/styles.xml
@@ -8,17 +8,4 @@
         <item name="colorAccent">@color/colorAccent</item>
     </style>
 
-    <style name="VoteButton" parent="Theme.MaterialComponents.Light.NoActionBar">
-        <item name="textAppearanceButton">@style/AppTextAppearance.Button</item>
-    </style>
-    <!--To fix rendering in preview -->
-    <style name="AppTextAppearance.Button" parent="TextAppearance.MaterialComponents.Button">
-        <item name="android:textAllCaps">true</item>
-    </style>
-
-    <style name="ShapeAppearanceOverlayExtended" parent="">
-        <item name="cornerFamily">rounded</item>
-        <item name="cornerSize">0dp</item>
-    </style>
-
 </resources>
